x86_64: limit maximum number of supported CPUs
With APIC/xAPIC interrupt controllers, there are only 256 APIC IDs
available. One is used for broadcasting, another is one is used for
IOAPIC/IOxAPIC, so only 254 CPUs at maximum may be supported.
Actually 254 is a large number for firecracker's usage cases.

This also fixes the panicking when starting a VM with 255 vCPUs.

Signed-off-by: Liu Jiang <gerry@linux.alibaba.com>