vmm: replace process::exit with libc::_exit
We did this because process::exit causes some extra syscalls to be
invoked, and we don't want to whitelist them. Using libc::_exit
will terminate the process more abruptly, but that's not an issue
in our case, because we don't create/use resources which are not
implicitly cleaned up by the OS on process termination.

Signed-off-by: Alexandru Agache <aagch@amazon.com>