kvm dirty log: do not assume mem_size
We were assuming that the guest memory size is a multiple of 256KB.
There is no need to do that and instead we can add an extra element in
the dirty page bitmap when required.
Renamed the function to `get_dirty_log` as this is the name of the ioctl
and all functions in KVM map to the ioctls that they implement wrappers
for.

Signed-off-by: Andreea Florescu <fandree@amazon.com>