vmm: run_control changes
Renamed run_control to run_event_loop, and changed it to return a
value represented by variants of the new enum EvenLoopExitReason. The
method only actually exits when VmmActionRequest event is triggered.
The run_control_action method is now invoked in the main "fc_vmm" thread
spawned by start_vmm_thread, and outside the Vmm struct. Essentially,
what happens is the event loop runs as before, until the next
VmmActionRequest is triggered. At that point the function returns,
then some other logic can handle the event and (usually) restart the
event loop. This is one of the first steps towards a more modular VMM,
where the logic that controls the Vmm struct resides in an outside
implementation that interacts with Vmm strictly through its interface.

Signed-off-by: Alexandru Agache <aagch@amazon.com>