added vsock definition in experimental yaml
Added firecracker-experimental.yaml in api_server/swagger.
This file is a copy of firecracker.yaml with an additional definiton
for the vsock API request.

The point of this file is to be used by third-party projects (like
Kata Containers) to automatically generate an API client that knows
how to send requests to the Firecracker API. The definition currently
lies in a different file and not in firecracker.yaml because vsock
is consider an experimental feature. Once the production ready vsock
is merged we will get rid of this experimental yaml.

For tracking purposes, this is the issue regarding switching to
vsock with UDS:
https://github.com/firecracker-microvm/firecracker/issues/650

In the new proposed vsock implementation, the API request for
configuring the vsock will also change.

Fixes https://github.com/firecracker-microvm/firecracker/issues/1085

Signed-off-by: Andreea Florescu <fandree@amazon.com>