MMIODeviceManager: bug fix
MMIODeviceManager::id_to_dev_info is a HashMap used to get the device
info starting from a device_id. The problem is that the key used by this
HashMap is only the device_id. If we have 2 devices of different types
that have the same device_id (e.g., a block device and a net device both
having device_id="1") id_to_dev_info will only contain the device info for
the last device that was added and the device info for the other device will
be lost. We add the device type to the key in order to fix this.

Signed-off-by: Serban Iorga <seriorga@amazon.com>