CI: separate microvm cleanup logic
Our CI offers the option of commenting out the
teardown function from testrun.sh so that the tests
do not recompile all the resources from test_session_path.
However, having the teardown commented will also leave all the
resources created for each test run. With this commit, those
resources will be deleted even when teardown is commented.
In the future, we could add an option to not do teardown if the
user wants so.

Signed-off-by: Diana Popa <dpopa@amazon.com>