devices/vsock: unix domain sockets backend
This is the first commit in the final season of the vsock saga. Building
on top of the vhost-less vsock device and the vsock connection state
machine, this brings the Unix Domain Sockets backend for vsock to
Firecracker.

The UDS backend-specific functionality resides mostly with the
`VsockMuxer`, in `devices/src/virtio/vsock/unix/muxer.rs` - a vsock
connection multiplexer, and a translator between guest-side AF_VSOCK and
host-side AF_UNIX streams.

Signed-off-by: Dan Horobeanu <dhr@amazon.com>