Removed dependency on time and chrono crates
Instead of using the time and chrono crates, we implemented some
functionality in the 'fc_util' crate. It uses the UNIX syscall
'clock_gettime', which is also what time and chrono used on Linux,
so no precision is lost. Where possible, we used 'Instant' from
'std::time' to measure elapsed time, as the time crate was deprecated
and it is at least as accurate. However, we could not use 'std::time'
everywhere as it doesn't support getting CPU clocks and system
wall time.

Fixes: https://github.com/firecracker-microvm/firecracker/issues/1123

Signed-off-by: George Pisaltu <gpl@amazon.com>