api_server: fix logic for  logging of API request
We would send the request to the VMM for processing
and only after that we would log that we received an
API request. This does not make sense if send_to_vmm
were to return error because you would want to see
that the server firstly received that API request
and only after that have the error message appear.

Signed-off-by: Diana Popa <dpopa@amazon.com>