GIC: Adding support for GICv2
We refactor the GIC-related code, to define a GICDevice Trait and then
create a GICv3 object that implements this Trait. This way, it is more
straightforward to introduce an implementations for different GIC
versions.

Moreover, we implement the Trait for GICv2 interrupt controllers.

At the moment, the creation of the GICDevice tries to create a version 3
device. If that fails, we fallback to GICv2.

Signed-off-by: Babis Chalios <babis.chalios@gmail.com>