tests: add details about the boot time tests
We should specify what kernel & root file system is used for running
the boot time integration test. This is not easy to guess from our
test code as you need to know which S3 images have the tag "boottime"
and then mapped them to the actual kernel+rootfs. For this purpose,
this is now documented in the high level documentation of the boot time
tests.

Signed-off-by: Andreea Florescu <fandree@amazon.com>