tests: fixed pylint errors - python 3.6.8
With the new version of python, the linter complains about quite a few
things.

1. Subprocess run should check the result of the command. In some
cases this is avoided on purpose because we expect the command to fail
with a particular status code. In these cases, the subprocess-run-check
is disabled. Also, when doing the cleanup, we don't need to check if the
cleanup is successful because we need to kill the process no matter
what.

2. Unused imports; this seems like a bug, so we just ignore pylint.

3. Unnecessary "else" after "continue" (no-else-continue); re-wrote the
if case.

4. Unnecessary "elif" after "break" (no-else-break); re-wrote the
if case.

Signed-off-by: Andreea Florescu <fandree@amazon.com>