virtio: group device status fields in module
Renamed the device status field so that there's a 1:1 mapping to
the Virtio specification. All of these fields are now moved to
a module called device_status. Added documentation to the module.
Also, changed the values from hex to decimal as this is how
they're defined in the specification.

Why? It should be easier to follow the specification if the names
of variables are a 1:1 mapping to the ones from the specification.

Signed-off-by: Andreea Florescu <fandree@amazon.com>