Remove `--seccomp-level` argument from jailer parameters
`--seccomp-level` command line argument, if passed to
jailer process, was only sent further to Firecracker.
This was a leftover from when the jailer was setting the
seccomp filters. Remove it from jailer arguments; now this
argument should be specified after `--` (end of command
options) in order to be forwarded to Firecracker process.

Signed-off-by: Laura Loghin <lauralg@amazon.com>