reworked checked_range_offset and get_host_address
It's been decided we're not going to upstream the checked_range_offset
and the modified version of get_host_address to vm-memory, and these
changes bring things in line with the future vm-memory integration.

checked_range_offset has been removed entirely, together with the
associated check in DescriptorChain::is_valid. The validity of memory
accesses is checked at the time of read/write when the device model
calls GuestMemory methods. Vsock devices are currently an exception,
but they do enforce their own validations via get_host_address (the
two parameter version, that won't get upstream to vm-memory). On that
front, we revert GuestMemory::get_host_address to the original, single
parameter version, and add a new one local to vsock/packet.rs that
implements the two-parameter behavior, to keep current validation
functionality. As we transition to vm-memory, we'll decide what
primitive we'll base the check on (seems like GuestMemory::do_in_region
will not be available anymore). Also, in the longer term we plan to
switch the Vsock implementation to do accesses via a GuestMemory
object, but that's still incipient at this point.

Signed-off-by: Alexandru Agache <aagch@amazon.com>