EventManager::register() doesn't wrap handler itself
'register()' shouldn't take ownership of the handler because it could
fail and on the Err path it would drop the handler.
Instead of taking ownership and wrapping it then returning it in an
Arc<Mutex>, just take an Arc<Mutex>. This way, the caller has maximum
flexibility and can choose to provide a simple clone and doesn't
lose the handler in an error case.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>