Fix field sizes vec for union (de)serialization
The functions that should calculate the sizes of the fields
available at some version for unions were not calling correctly
`std::mem::size_of::<T>()`. Because of that, `generate_field_finder()`
returned the union field with the smallest size instead of the largest
one when serializing or deserializing the union.

Signed-off-by: Laura Loghin <lauralg@amazon.com>