aarch64: harden core register id calculation
Replace old macro that would hardcode the field of
the structure in which to calculate offsets to a new
one that lets you compute the offset outside of it.
Also, thoroughly calculate the offset by also including the
offset of the user_pt_regs field inside the kvm_regs structure.

Signed-off-by: Diana Popa <dpopa@amazon.com>