vmm: implement device manager (de)serialization
Move device map crawling and serialization from Vmm to the
device manager.

The device manager implements 'save()' which
creates the previously defined 'DeviceStates' containing
the devices' states plus mmio bus and irq attributes.

The device manager implements 'restore()' which takes
a 'DeviceStates' structure, restores the devices and
registers them using the original mmio bus and irq settings.

Ideally the device-manager would be device implementation agnostic,
but Persist and Versionize don't allow dynamic dispatch.
The current organization of serialized data does however allow such
a design, so the present state is a two-way-door and if in the future
we invest here, we might be able to get the implementation there with
no disruption on the user.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>