tests: fix net rate limiters test
This commit makes the test more reliable by using a fixed TCP
window size and disabling TCP Nagle.

It is however still not ideal. The problem is that it relies on iperf3
which measures transport protocol bandwidth and compares that to
raw wire bandwidth. To make matters worse, iperf3 and the guest
kernel vary transmission parameters in their attempt to maximize
TCP throughput. For short test durations like ours, resulting numbers
show jitter.

Another thing we could try is to move to UDP iperf3 tests, but that's
also not trivial because UDP packet size is not configurable in iperf3
and it uses small packets making the total difference between raw bytes
and protocol bytes even larger.

An accurate test should use a low level tool to send an exact number
of raw bytes and measure that against time taken.

Signed-off-by: Adrian Catangiu <acatan@amazon.com>